+++ tests/files/src/0886_void_argument.php
@@ -1,0 +1,2 @@
+<?php
+echo json_encode(debug_print_backtrace());